Add cpp ext lock file check during ORTModule init
